update to block and warning fixes

+++ cryptoTools/Crypto/blake2/c/blake2bp-ref.c
@@ -51,7 +51,7 @@ static int blake2bp_init_leaf( blake2b_state *S, size_t outlen, size_t keylen, u
   P->fanout = PARALLELISM_DEGREE;
   P->depth = 2;
   store32( &P->leaf_length, 0 );
-  store32( &P->node_offset, offset );
+  store32( &P->node_offset, (uint32_t)offset );
   store32( &P->xof_length, 0 );
   P->node_depth = 0;
   P->inner_length = BLAKE2B_OUTBYTES;
